validator_structure: new structure
