+++ microdata_tools/__init__.py
@@ -1,4 +1,14 @@
-from microdata_tools.package_dataset import package_dataset
-from microdata_tools.unpackage_dataset import unpackage_dataset
+from microdata_tools.packaging import package_dataset
+from microdata_tools.packaging import unpackage_dataset
+from microdata_tools.validation import (
+    validate_dataset, validate_metadata, inline_metadata
+)
 
-__all__ = ["package_dataset", "unpackage_dataset"]
+
+__all__ = [
+    "package_dataset",
+    "unpackage_dataset",
+    "validate_dataset",
+    "validate_metadata",
+    "inline_metadata"
+]
