+++ microdata_tools/packaging/_utils.py
